Adds some tests and moves nockify to a different file
